Moved and edited 3D scene example
